requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.5
django-import-export==2.5.0
django-jazzmin==2.4.4
django-rest-framework==0.1.0
django-sitemaps==1.1.1
djangorestframework==3.12.2
djongo==1.3.4
drf-yasg==1.20.0
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
lxml==4.6.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.9
Pillow==8.1.0
pymongo==3.11.3
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.2.4
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
